Factor out ScopeGroup from BindingsSection
